// src/types/enums.ts
export enum EnScreens {
  TASK_LIST = 'TaskList',
  TASK_EDIT = 'TaskEdit',
  SEARCH = 'Search',
}
export enum EnSortType {
  PRIORITY = 'priority',
  DEDLINE = 'dedline',
  STATUS = 'status',
}
export enum EnFilterType {
  WORK = 'Work',
  PERSONAL = 'Personal',
  ALL = 'All',
}
export enum EnFilterValues {
  STATUS = 'status',
  PRIORITY = 'priority',
  CATEGORY = 'category',
}
export enum EnMessagePreset {
  SUCCESS = 'SUCCESS',
  INFO = 'INFO',
  ERROR = 'ERROR',
  LOADING = 'LOADING',
}